chore: move getCololrs to utilities
