+++ src/admin/utilities/getColors.ts
